be/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
orjson==3.10.7
scikit-learn==1.5.2
joblib==1.4.2
numpy==2.1.2
pandas==2.2.3
httpx==0.27.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
groq==0.4.2
python-dotenv==1.0.0
slowapi==0.1.9
python-whois==0.9.4
dnspython==2.6.1
aiohttp==3.10.5

